Improve acceptance testing & fix bug when project id is empty (#12)
ENHANCEMENT:

- Add test scenario and improve the testing framework
- Improve continuous deployment workflow, integration on a private cluster

BUGFIX:

- Fix unexpected error when project id is empty

+++ ucloud/services/udb/schemas/models.py
@@ -151,7 +151,7 @@ class UDBInstancePriceSetSchema(schema.ResponseSchema):
 
     fields = {
         "ChargeType": fields.Str(required=False, load_from="ChargeType"),
-        "Price": fields.Float(required=False, load_from="Price"),
+        "Price": fields.Int(required=False, load_from="Price"),
     }
